requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# HTTP 请求
requests>=2.31.0
aiohttp>=3.9.0

# 日期处理
pendulum>=3.0.0

# GitHub API
PyGithub>=2.1.1

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理
duckdb>=0.9.0

# 中文转换
zhconv>=1.4.3

# 开发工具 (可选)
pytest>=7.4.0
black>=23.0.0
